# docs/swagger.yaml
basePath: /api/v1
definitions:
  Kurajj_internal_models.Address:
    properties:
      city:
        type: string
      district:
        type: string
      homeLocation:
        type: string
      region:
        type: string
    type: object
...
  Kurajj_internal_models.CreationResponse:
    properties:
      id:
        type: integer
    type: object
  Kurajj_internal_models.ErrResponse:
    properties:
      error:
        type: string
    type: object
...
  Kurajj_internal_models.Ids:
    properties:
      IDs:
        items:
          type: integer
        type: array
    type: object
...
  Kurajj_internal_models.NotificationResponse:
    properties:
      createdAt:
        type: string
      eventID:
        type: integer
      eventTitle:
        type: string
      id:
        type: integer
      isRead:
        type: boolean
      newStatus:
        type: string
      text:
        type: string
    type: object
...
  Kurajj_internal_models.RefreshTokenInput:
    properties:
      refreshToken:
        type: string
    type: object
  Kurajj_internal_models.SearchValueResponse:
    properties:
      id:
        type: integer
      title:
        type: string
      values:
        items:
          type: string
        type: array
    type: object
  Kurajj_internal_models.SignInEntity:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  Kurajj_internal_models.SignUpUser:
    properties:
      address:
        $ref: '#/definitions/Kurajj_internal_models.Address'
      companyName:
        type: string
      email:
        type: string
      fileBytes:
        items:
          type: integer
        type: array
      fileType:
        type: string
      firstName:
        type: string
      password:
        type: string
      secondName:
        type: string
      telephone:
        type: string
    type: object
  Kurajj_internal_models.SignedInUser:
    properties:
      address:
        $ref: '#/definitions/Kurajj_internal_models.Address'
      companyName:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      profileImageURL:
        type: string
      proposalEventSearchValues:
        items:
          $ref: '#/definitions/Kurajj_internal_models.SearchValueResponse'
        type: array
      refreshToken:
        type: string
      secondName:
        type: string
      telephone:
        type: string
      token:
        type: string
      transactionNotifications:
        items:
          $ref: '#/definitions/Kurajj_internal_models.NotificationResponse'
        type: array
    type: object
...
  Kurajj_internal_models.TokensResponse:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: mykha
    name: API Support
    url: http://www.swagger.io/support
  description: Kurajj Charity Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger Cor Charity Platform
  version: "1.0"
paths:
...
  /api/read-notifications:
    put:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Kurajj_internal_models.Ids'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Changes notifications status to read=true
      tags:
      - User
  /api/refresh-user-data:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Kurajj_internal_models.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Kurajj_internal_models.SignedInUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Return user data
      tags:
      - User
...
  /auth/confirm/{email}:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Updates user's status to 'activated'.
      tags:
      - Auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Kurajj_internal_models.RefreshTokenInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Kurajj_internal_models.TokensResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Update access token expiration date
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Kurajj_internal_models.SignInEntity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Kurajj_internal_models.SignedInUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Signs In a user
      tags:
      - Auth
...
  /auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/Kurajj_internal_models.SignUpUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Kurajj_internal_models.CreationResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Kurajj_internal_models.ErrResponse'
      summary: Signs Up new user
      tags:
      - Auth
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"